+++ src/Main.java
@@ -1,5 +1,5 @@
 //package contacts;
-//import contacts.*;
+import contacts.*;
 public class Main {
     public static void main(String[] args) {
         java.util.Scanner input = new java.util.Scanner(System.in);
